scripts/requirements.txt
transformers
torch
langdetect
pillow